requirements.txt
playwright
pyyaml
requests
httpx
numpy
pandas
matplotlib

# for debugging and tracing
snoop
# for cli parsing
fire